Dev.amr

+++ src/components/horizontal-scroll.tsx
@@ -1,11 +1,10 @@
 import { FC, ReactNode, useRef, useState, useEffect } from "react";
-import Image from "next/image";
-import { Button } from "./ui/button/button";
-import Link from "next/link";
 import { gsap } from "gsap";
 import { ScrollTrigger } from "gsap/dist/ScrollTrigger";
-import { useLocale, useTranslations } from "next-intl";
+import { useLocale } from "next-intl";
 import styles from './styles.module.scss';
+import { useMediaQuery } from "@/hooks/use-media-query";
+import { Card } from "./ui/service-card";
 
 interface Props {
   cards: {
@@ -23,7 +22,10 @@ export default function HorizontalScrollCarousel() {
   const triggerRef = useRef<HTMLDivElement | null>(null);
   const [currentCard, setCurrentCard] = useState(0);
 
+
   let isRTL = useLocale() === 'ar';
+
+
 
 
 
@@ -53,10 +55,11 @@ export default function HorizontalScrollCarousel() {
     });
 
     return () => {
-      pin.kill();
+      pin && pin.kill();
     };
   }, [isRTL]);
   return (
+
     <section className={styles.scrollSectionOuter}>
       <div ref={triggerRef} className={styles.scrollContainer}>
         <div ref={sectionRef} className={styles.scrollSectionInner}>
@@ -69,54 +72,9 @@ export default function HorizontalScrollCarousel() {
         <Indicator total={4} current={currentCard} />
       </div>
     </section>
+
   );
 }
-
-const Card = ({ card }: { card: number }) => {
-  const t = useTranslations(`OurServices.cards.${card}`);
-  return (
-    <div className="flex items-center w-full h-full p-8">
-      <div className="h-full flex justify-center  w-5/12">
-        <Image
-          src={t('image')}
-          alt={t('title')}
-          width={500}
-          height={500}
-          unoptimized
-          className="object-contain h-full w-full"
-        />
-      </div>
-      <div className="flex items-center justify-center h-full flex-1 ">
-        <div className="text-center w-3/4 flex flex-col gap-4">
-          <span className="ltr:tracking-widest ltr:font-mono dark:text-red-400 font-bold text-3xl text-red-700">
-            {
-              t('subtitle')
-            }
-          </span>
-          <h3 className="text-9xl font-bold ltr:tracking-tighter rtl:leading-tight">
-            {
-              t('title')
-            }
-          </h3>
-          <p className="text-2xl leading-relaxed">
-            {
-              t('description')
-            }
-          </p>
-          <Button variant='ringHover' className="text-3xl py-8" size='lg'>
-            <Link href={
-              t('link')
-            }>
-              {
-                t('btnText')
-              }
-            </Link>
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 const Indicator = ({ total, current }: {
   total: number;
